feat: Create Rooms
Completed the logic to register a new room and send the confirmation email

Once the room is created, a function is called to assing this to the corresponding user.
In this function, a call for background job to send the email confirmation is executed,
with the user.id and room.id params.

The email confirmation job sets both the user and room, and finally the call to send
the email is executed, with user instance and room as params.